+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.15.0b2 (Unreleased)
+## 1.15.0b2 (2023-10-12)
 
 ### Features Added
 
@@ -16,7 +16,11 @@
 
 - `ManagedIdentityCredential` will now correctly retry when the instance metadata endpoint returns a 410 response.  ([#32200](https://github.com/Azure/azure-sdk-for-python/pull/32200))
 
-### Other Changes
+## 1.14.1 (2023-10-09)
+
+### Bugs Fixed
+
+- Bug fixes for developer credentials
 
 ## 1.15.0b1 (2023-09-12)